pyproject.toml
[project]
name = "networks-and-system-sec-portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "binwalk>=2.1.0",
    "capstone>=5.0.6",
    "cryptography>=46.0.2",
    "keystone-engine>=0.9.2",
    "lief>=0.17.1",
    "pillow>=12.0.0",
    "pyotp>=2.9.0",
    "python-nmap>=0.7.1",
    "qrcode>=8.2",
    "scapy>=2.6.1",
    "yara>=1.7.7",
]
